# bff/openapi-schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BeTrace Behavioral Assurance API",
    "version": "1.0.0",
    "description": "Real-time Behavioral Assurance System for OpenTelemetry data analysis and signal management",
    "contact": {
      "name": "BeTrace Team",
      "email": "info@betrace.scoble.me"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://localhost:8443",
      "description": "HTTPS Server"
    },
    {
      "url": "http://localhost:8080",
      "description": "HTTP Server"
    }
  ],
...
  "paths": {
    "/api/v1/rules": {
      "get": {
        "tags": ["Rules"],
        "summary": "List all active rules",
        "description": "Retrieve all active rules in the system",
        "operationId": "getAllRules",
        "responses": {
          "200": {
            "description": "List of active rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rules"],
        "summary": "Create a new rule",
        "description": "Create a new behavioral assurance rule",
        "operationId": "createRule",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid rule data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - rule already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rules/{ruleId}/{version}": {
      "delete": {
        "tags": ["Rules"],
        "summary": "Delete a rule",
        "description": "Logically delete a specific rule version",
        "operationId": "deleteRule",
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The rule identifier"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The rule version"
          }
        ],
        "responses": {
          "200": {
            "description": "Rule deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Rule not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Rule": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "description": "Unique identifier for the rule"
          },
          "ruleVersion": {
            "type": "string",
            "description": "Version of the rule"
          },
          "ognlExpression": {
            "type": "string",
            "description": "OGNL expression defining the rule logic"
          },
          "description": {
            "type": "string",
            "description": "Human-readable description of the rule"
          },
          "activeStatus": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "DELETED"],
            "description": "Current status of the rule"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "description": "When the rule was created"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "description": "When the rule was last modified"
          }
        },
        "required": ["ruleId", "ruleVersion", "ognlExpression", "activeStatus"]
      },
      "CreateRuleRequest": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "description": "Unique identifier for the rule (lowercase, alphanumeric with hyphens)"
          },
          "ruleVersion": {
            "type": "string",
            "description": "Version of the rule (auto-assigned if not provided)"
          },
          "ognlCondition": {
            "type": "string",
            "description": "OGNL expression defining the rule logic"
          },
          "description": {
            "type": "string",
            "description": "Human-readable description of the rule"
          }
        },
        "required": ["ruleId", "ognlCondition"]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          }
        },
        "required": ["message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error", "message"]
      }
    }
  }
}